Time: 0 ms (100.00%), Space: 7.9 MB (10.53%) - LeetHub

+++ 1614-maximum-nesting-depth-of-the-parentheses/1614-maximum-nesting-depth-of-the-parentheses.cpp
@@ -7,8 +7,8 @@ public:
         for(int i=0;i<n;i++){
             if(s[i]=='('){
                 st.push('(');
-                int n=st.size();
-                mxdp=max(mxdp,n);
+                long n=st.size();
+                mxdp=max(mxdp,(int)n);
             }
             else if(s[i]==')'){
                 st.pop();
